# src/RESTApi/managementApi/tsOpenApi/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"BoxCalendarLink": {
				"properties": {
					"boxId": {
						"type": "string"
					},
					"calendarId": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"boxId",
					"calendarId",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CalendarUpdateForm": {
				"properties": {
					"calendarId": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"calendarId",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": true
			},
			"MsTokenPair": {
				"properties": {
					"accessToken": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CalendarConfigurationForm": {
				"properties": {
					"boxId": {
						"type": "string"
					},
					"calendarId": {
						"type": "string"
					},
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"boxId",
					"calendarId",
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Calendar Service Management API",
		"version": "0.5.0"
	},
	"openapi": "3.0.0",
	"paths": {
		"/configurations/all": {
			"get": {
				"operationId": "GetAllCalendarLinks",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/BoxCalendarLink"
									},
									"type": "array"
								}
							}
						},
						"description": "Ok"
					}
				},
				"summary": "Retrieve all box entries in the database.",
				"security": [],
				"parameters": []
			}
		},
		"/configurations/{boxId}": {
			"get": {
				"operationId": "GetCalendarLink",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BoxCalendarLink"
								}
							}
						},
						"description": "Ok"
					}
				},
				"summary": "retrieve a specific boxCalendarLink from the database.",
				"security": [],
				"parameters": [
					{
						"description": "the box of which you wish to retrieve the database entry of.",
						"in": "path",
						"name": "boxId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "PutCalendarLink",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						},
						"description": "Ok"
					}
				},
				"summary": "Add or update the calendar link of a specific box.",
				"security": [],
				"parameters": [
					{
						"description": "Box that should be updated.",
						"in": "path",
						"name": "boxId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CalendarUpdateForm"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCalendarLink",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						},
						"description": "Ok"
					}
				},
				"summary": "Deletes a given boxId from the collection.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "boxId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/MicrosoftTokens": {
			"get": {
				"operationId": "MicrosoftTokens",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MsTokenPair"
								}
							}
						},
						"description": "Ok"
					}
				},
				"summary": "Exchanges the code provided by Microsoft for AccessTokens and RefreshTokens.",
				"security": [],
				"parameters": [
					{
						"description": "OAuth2 code provided by the msAuthentication flow.",
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "client_credentail",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tenant",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/saveCalendarConfiguration": {
			"post": {
				"operationId": "SaveCalendarConfiguration",
				"responses": {
					"204": {
						"content": {
							"application/json": {}
						},
						"description": "No content"
					}
				},
				"summary": "Save a new Calendar configuration in to the database.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CalendarConfigurationForm"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:30662/api/v1.0"
		}
	]
}